requirements.txt
feedparser
trafilatura
jinja2
PyYAML
requests
lxml
